[191][cosmos-proxy] Fix issues raised during code review
